Finally got the test case to fail with the incorrect values.
Had multiple problems such as ErrorAs (I made wrong assumption), and I was a bit sloppy with decomposing the Cause, which I've changed to use errors.Unwrap().

I'll continue with the rest of the existing test output, converting them over while also finishing up the rest of the testTypes. As a comment, unrelated to commit, maybe I should do this with the traps as well before continuing with div_s.i64.