Finished converting empty type
The PopulateImage() in image.go was getting unruly, so extracted each type to different functions for ones that are converted.

Found missed case for empty and coverage squeaked back up to 94.1% for ./pkg since it had dropped to 93%.

The ApplyCause(), MarshalJSON(), and Error() are not covered. I'm commenting out MarshalJSON() and ApplyCause() for now. Error(), I'll leave it intact for now. This bumped it up to 95.3%