Wrap up vm_tests primary refactor
We have 95% appoximate coverage after refactor. I included as part of it the error struct for VM initialization.

This one, I took a different approach and did the error messages as a map. If I end up liking it, I'll refactor Image later.

What I did not do is refactor the errors related to Traps into the table structure. That'll get handled during the specific ticket later.